backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.11.3
pydantic-settings==2.7.1

sqlalchemy==2.0.39
aiosqlite==0.19.0
asyncpg==0.30.0

httpx==0.26.0
requests==2.31.0

zhipuai>=2.0.0
boto3==1.35.55

python-dateutil==2.8.2
python-dotenv==1.0.0

reportlab==4.1.0
markdown==3.5.1
python-pptx==0.6.23
python-docx==1.1.2
pypdf==5.1.0
openpyxl==3.1.5
